7.0 Garrett Prototype 5
Parental controls screen time limit (should be) fully complete and working optimally. shock-users and shock-initial-setup need a small revision to prevent attempting to create a user if the username does not meet the character requirements. shock-auto-updates*, shock-os-update-manager, and shock-release upgrader will need to be tested, and probably need a few revisions before they are production-ready.