7.0 Garrett Prototype 6
shock-users and shock-initial-setup (both programs create can create a user) now properly check to see if the username matches the regualar expression configured via the NAME_REGEX variable. However, they do not properly check to make sure that both passwords match each other when varifying a password (when creating a user). Passwords with certain symbols won't work (probably the way the passwords are quoted in the boolean needs to be changed). Afterwards, shock-auto-updates*, shock-os-update-manager and shock-version-upgrader need to be tested (and likely revised) before they are production ready.