7.0 Garrett Prototype 7
shock-initial-setup turned out to be fine for password matching. shock-users now uses the "`" character as the delimiter instead of "|" as people are less likely to have usernames/fullnames/passwords with that character in it. shock-users also prevents entering the "`" character in any field to avoid glitches. shock-auto-updates*, shock-os-update-manager and shock-release-upgrader need to be tested, and will probably require revisions, before they are production-ready.