7.0 Garrett Prototype 10 (RC1)
shock-release-upgrader polished and ready for release. shock-os-update-manager scrapped and replaced with a much simpler, but fully working version. The full version of shock-os-update-manager that was intended for this release will have to wait until 8.0 Holly. Shock OS Imager (shock-os-imager) has been scrapped, and will also have to wait until 8.0 Holly. shock-slideshow-wallpaper recieved a small patch to be able to refresh the wallpapers directory automatically, without requiring a reboot or re-applying the wallpaper preferences. This means that shock-slideshow-wallpaper can now detect when a wallpaper has been added to, or removed from, the slideshow, thus avoiding glitches.
	renamed:    shock/shock-release-upgrade.svg -> shock/shock-release-upgrader.svg
	deleted:    usr/bin/shock-os-imager
	modified:   usr/bin/shock-os-update-manager
	modified:   usr/bin/shock-release-upgrader
	modified:   usr/bin/shock-slideshow-wallpaper
	deleted:    usr/share/applications/shock-os-imager.desktop
	deleted:    usr/share/shock-os-imager/icon.svg