7.0 Garrett Prototype 15 (RC6)
shock-users recieved a slight but needed fix to better grep the list of logged in users when attempting to remove a user. This ensures accurated detection of who is logged into the MATE Desktop Environment and gets rid of the ridiculous restriction of not allowing usernames to consist of 'x-session-manager'.
	modified:   usr/bin/shock-users