8.0 Holly Prototype 3
File /shock/setupfile now deleted after initial setup is completed. LICENSE moved to /shock. shock-auth created to replace the Shock Authentication Element (formerly password-detector). Rather than being embedded into apps, as SAE was, shock-auth is a separate app that is launched with a command to run as an argument. If the password is entered correctly, the command will be executed with the user that opened the app as the first argument. This is only intended to be used with other Shock OS scripts, not designed for third party apps. shock-auth uses encryption and ranomization, as well as clearing the password from the memory as soon as it is no longer needed, to provide security that is far superior to SAE. Shock OS apps have been updated to comply with it. shock-backup-and-restore added with experimental backup personal files functionality currently, more functionality coming later.
	modified:   DEBIAN/postinst
	new file:   shock/Easter Egg.png
	renamed:    LICENSE -> shock/LICENSE
	modified:   shock/applications/users.desktop
	new file:   usr/bin/shock-auth
	modified:   usr/bin/shock-auto-updates-setup
	new file:   usr/bin/shock-backup-and-restore
	modified:   usr/bin/shock-os-imager
	modified:   usr/bin/shock-os-update-manager
	modified:   usr/bin/shock-panel-layout-changer
	modified:   usr/bin/shock-parental-controls
	modified:   usr/bin/shock-release-upgrader
	modified:   usr/bin/shock-startup
	modified:   usr/bin/shock-users