8.0 Holly Prototype 6
shockware-center and shockware-center-show-package added. Yes, this is a software manager for Shock OS, and it is intended to replace pi-packge. pi-package has not been removed yet, but it will be when shockware-center is fully ready (which it almost is). shock-os-update-manager will also be completely revamped. The new version of shock-os-update-manager will be based off of the new shockware-center. A new .deb package installer *might* also be created to replace the one provided by pi-package, or we will bring back GDebi.
	modified:   DEBIAN/control
	modified:   DEBIAN/postinst
	modified:   usr/bin/shock-auto-updates-setup
	modified:   usr/bin/shock-backup-and-restore
	modified:   usr/bin/shock-fonts
	modified:   usr/bin/shock-initial-setup
	modified:   usr/bin/shock-os-imager
	modified:   usr/bin/shock-panel-layout-changer
	modified:   usr/bin/shock-parental-controls
	modified:   usr/bin/shock-release-upgrader
	modified:   usr/bin/shock-startup
	modified:   usr/bin/shock-users
	new file:   usr/bin/shockware-center
	new file:   usr/bin/shockware-center-show-package
	new file:   usr/share/shockware-center/apt16.png
	new file:   usr/share/shockware-center/apt32.png
	new file:   usr/share/shockware-center/clean.png
	new file:   usr/share/shockware-center/flatpak16.png
	new file:   usr/share/shockware-center/flatpak32.png
	new file:   usr/share/shockware-center/shockware-center.svg