8.0 Holly Prototype 7
shockware-center recieved added support for detecting if flatpak is installed, and only including flatpaks in results if flatpak is installed and enabled. Flatpaks is now pre-installed (as a recommend in the control file, so it can be removed if the user wishes), and FlatHub is preconfigured in the postinst script.
	modified:   DEBIAN/control
	modified:   DEBIAN/postinst
	modified:   usr/bin/shockware-center