8.0 Holly Prototype 11
FlatHub is now set up in shock-startup to avoid errors when building Shock OS. shockware-center is now no longer started as root, but uses pkexec to gain root privilages when installing/removing with apt or dpkg. This allows Flatpaks to be installed as the normal user. shockware-center also has a new icon.
	modified:   DEBIAN/postinst
	modified:   usr/bin/shock-startup
	modified:   usr/bin/shockware-center
	modified:   usr/share/shockware-center/icon.svg